requirements.txt
rdkit-pypi==2022.9.5
tqdm==4.65.0
pandas==1.5.3